+++ src/main/store/legacy-types.ts
@@ -197,6 +197,7 @@ export type v1_6_0_KeyboardSettings = {
   fadeOutDuration: number;
   textOpacity: number;
   displaySustained: boolean;
+  wrap: boolean;
   sizes: {
     radius: number;
     height: number;
@@ -207,6 +208,7 @@ export type v1_6_0_KeyboardSettings = {
     white: string | null;
     black: string | null;
     played: string | null;
+    wrapped: string | null;
     sustained: string | null;
   };
 };