+++ src/renderer/components/PianoKeyboard/classic/Labels.tsx
@@ -15,6 +15,7 @@ const Board: React.FC<Props> = ({ keys, sizes }) => (
   <g className={styles.labels} transform="translate(0,0)">
     {keys.labels.map(({ note, labelOffset }) => (
       <text
+        key={note.midi}
         className={classNames([styles.pianoLabel, `label-${note.midi}`])}
         x={labelOffset}
         y={sizes.LABEL_HEIGHT - sizes.LABEL_OFFSET}