+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Features
 - **Chord Display:** Fully customizable Piano (sizes, colors and labels above keys)
 - **Chord Display:** disable sustain pedal for detection or display
+- **Chord Display:** wrap keyboard option
 
 # 1.5.1 (2023-11-21)